02-Quantum-Chemistry-VQE/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "quantum-chemistry-vqe"
version = "0.3.0"
description = "Modular VQE PES benchmarking platform for quantum chemistry."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "DEVADATH H K" }]
dependencies = [
  "numpy>=1.26",
  "scipy>=1.12",
  "matplotlib>=3.8",
  "pyyaml>=6.0",
  "qiskit>=1.0",
  "qiskit-nature>=0.7",
  "qiskit-algorithms>=0.3",
  "qiskit-ibm-runtime>=0.23",
  "qiskit-aer>=0.14",
  "pydantic>=2.7",
  "ipywidgets>=8.1",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "jupyter>=1.0"]
